SeqSign into new app. Changes into property groups, scoring, sign calc.
