+++ common/definitions.py
@@ -141,27 +141,27 @@ AMINO_ACID_GROUPS = OrderedDict([
     ('Y_-', ('Y')),
     ('W_-', ('W')),
     ('_', ('-', '_')),
-    ('-G', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'P', '-', '_')),
-    ('-A', ('L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-C', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'D', 'G', 'P', '-', '_')),
-    ('-S', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-T', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-V', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-P', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', '-', '_')),
-    ('-D', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'G', 'P', '-', '_')),
-    ('-I', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-L', ('A', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-N', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-M', ('A', 'L', 'R', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-E', ('A', 'L', 'R', 'M', 'K', 'Q', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-Q', ('A', 'L', 'R', 'M', 'K', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-H', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-K', ('A', 'L', 'R', 'M', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-F', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-R', ('A', 'L', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-Y', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('-W', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
-    ('--', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P')),
+    # ('-G', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'P', '-', '_')),
+    # ('-A', ('L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-C', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'D', 'G', 'P', '-', '_')),
+    # ('-S', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-T', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-V', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-P', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', '-', '_')),
+    # ('-D', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'G', 'P', '-', '_')),
+    # ('-I', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-L', ('A', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-N', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-M', ('A', 'L', 'R', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-E', ('A', 'L', 'R', 'M', 'K', 'Q', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-Q', ('A', 'L', 'R', 'M', 'K', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-H', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-K', ('A', 'L', 'R', 'M', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-F', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-R', ('A', 'L', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-Y', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('-W', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P', '-', '_')),
+    # ('--', ('A', 'L', 'R', 'M', 'K', 'Q', 'E', 'I', 'W', 'S', 'Y', 'F', 'V', 'H', 'N', 'T', 'C', 'D', 'G', 'P')),
 ])
 
 AMINO_ACID_GROUP_NAMES = OrderedDict([
@@ -222,27 +222,27 @@ AMINO_ACID_GROUP_NAMES = OrderedDict([
     ('Y_-', 'Tyr'),
     ('W_-', 'Trp'),
     ('_', 'Gap'),
-    ('-G', 'Not Gly'),
-    ('-A', 'Not Ala'),
-    ('-C', 'Not Cys'),
-    ('-S', 'Not Ser'),
-    ('-T', 'Not Thr'),
-    ('-V', 'Not Val'),
-    ('-P', 'Not Pro'),
-    ('-D', 'Not Asp'),
-    ('-I', 'Not Ile'),
-    ('-L', 'Not Leu'),
-    ('-N', 'Not Asn'),
-    ('-M', 'Not Met'),
-    ('-E', 'Not Glu'),
-    ('-Q', 'Not Gln'),
-    ('-H', 'Not His'),
-    ('-K', 'Not Lys'),
-    ('-F', 'Not Phe'),
-    ('-R', 'Not Arg'),
-    ('-Y', 'Not Tyr'),
-    ('-W', 'Not Trp'),
-    ('--', 'Not gap'),
+    # ('-G', 'Not Gly'),
+    # ('-A', 'Not Ala'),
+    # ('-C', 'Not Cys'),
+    # ('-S', 'Not Ser'),
+    # ('-T', 'Not Thr'),
+    # ('-V', 'Not Val'),
+    # ('-P', 'Not Pro'),
+    # ('-D', 'Not Asp'),
+    # ('-I', 'Not Ile'),
+    # ('-L', 'Not Leu'),
+    # ('-N', 'Not Asn'),
+    # ('-M', 'Not Met'),
+    # ('-E', 'Not Glu'),
+    # ('-Q', 'Not Gln'),
+    # ('-H', 'Not His'),
+    # ('-K', 'Not Lys'),
+    # ('-F', 'Not Phe'),
+    # ('-R', 'Not Arg'),
+    # ('-Y', 'Not Tyr'),
+    # ('-W', 'Not Trp'),
+    # ('--', 'Not gap'),
 ])
 AMINO_ACID_GROUPS_OLD = OrderedDict([
         ('hp',     ('A', 'C', 'F', 'I', 'L', 'M', 'P', 'V', 'W', 'Y')),
