+++ home/templates/home/mainmenu_gpcr.html
@@ -51,7 +51,7 @@
                 </button>
                 </a></li>
                 <li class="divider"></li>
-                <!-- SEQUENCEAS -->
+                <!-- SEQUENCES -->
                 <li class="nav-header disabled text-left"><a>SEQUENCES
 
                         <!-- <button type="button" class="pull-right btn btn-xs btn-default">
@@ -100,7 +100,7 @@
                         <li><a href="/alignment/blastsearch">Similarity search-BLAST</a></li>
                         <li><a href="/similaritymatrix/targetselection">Similarity matrix</a></li>
                         <li class="divider"></li>
-                        <li><a href="/alignment/positivegroupselection">Sequence signature</a></li>
+                        <li><a href="/seqsign">Sequence signature</a></li>
                     </ul>
                 </li>
