+++ seqsign/templates/sequence_signature.html
@@ -1,5 +1,7 @@
-{% extends "home/base.html" %} {% load staticfiles %} {% block addon_css %}
-<link href="{% static 'home/css/alignment.css' %}" rel="stylesheet">
+{% extends "home/base.html" %}
+{% load staticfiles %}
+{% block addon_css %}
+    <link href="{% static 'home/css/alignment.css' %}" rel="stylesheet">
 <!-- <style>
 .axis path,
 .axis line {
@@ -20,7 +22,7 @@
         <td>
             <form style="display: inline;">
                 <div class="input-group">
-                    <span class="input-group-addon" id="basic-addon3">Signature fraquency cutoff: </span>
+                    <span class="input-group-addon" id="basic-addon3">Signature score cut-off: </span>
                     <input id="cutoff-val" type="text" class="form-control" size=15 placeholder="40">
                     <div class="input-group-btn">
                         <button id="cutoff-apply" class="btn btn-outline-secondary" type="button"><i class="glyphicon glyphicon-filter"></i></button>
@@ -83,13 +85,13 @@
             <div class="ali-first-col-div">
                 <table>
                     <tr>
-                        <td class="ali-td ali-td-first-col">{% include 'sequence_signature/signature_left_panel.html' %}</td>
+                        <td class="ali-td ali-td-first-col">{% include 'signature_left_panel.html' %}</td>
                     </tr>
                     <tr>
-                        <td class="ali-td ali-td-first-col">{% include 'sequence_signature/signature_aln_pos_left_panel.html' %}</td>
+                        <td class="ali-td ali-td-first-col">{% include 'signature_aln_pos_left_panel.html' %}</td>
                     </tr>
                     <tr>
-                        <td class="ali-td ali-td-first-col">{% include 'sequence_signature/signature_aln_neg_left_panel.html' %}</td>
+                        <td class="ali-td ali-td-first-col">{% include 'signature_aln_neg_left_panel.html' %}</td>
                     </tr>
                 </table>
             </div>
@@ -106,13 +108,13 @@
                 <div class="dynamic-div">
                     <table>
                         <tr>
-                            <td>{% include 'sequence_signature/signature_right_panel.html' %}</td>
+                            <td>{% include 'signature_right_panel.html' %}</td>
                         </tr>
                         <tr>
-                            <td>{% include 'sequence_signature/signature_aln_pos_right_panel.html' %}</td>
+                            <td>{% include 'signature_aln_pos_right_panel.html' %}</td>
                         </tr>
                         <tr>
-                            <td>{% include 'sequence_signature/signature_aln_neg_right_panel.html' %}</td>
+                            <td>{% include 'signature_aln_neg_right_panel.html' %}</td>
                         </tr>
                     </table>
                 </div>
@@ -126,28 +128,27 @@
 <!-- sequence alignment ends -->
 {% endblock %}
 {% block addon_js %}
-<!-- <script src="https://d3js.org/d3.v4.min.js"></script> -->
-<script src="{% static 'home/js/alignment.js' %}"></script>
 
-<!-- <script src="{% static 'home/js/scatter_plot.js' %}"></script> -->
-<script src="{% static 'home/js/saveSvgAsPng.js' %}"></script>
-<script type="text/javascript" charset="utf-8">
-    $('cutoff-val').keydown( function (event) {
-        if (event.keyCode == 13) {
-            var cutoff = parseInt($('#cutoff-val').val());
-            ApplyCutoff(cutoff);
-        };
-    });
-    $('#score-button').click( function () {
-        var cutoff = $('#cutoff-val').val();
-        if (cutoff == '') {
-            cutoff = 40;
-        }
-        window.location.href = 'render_signature_match_scores/'+cutoff;
-    });
-    //$(window).load(function () {
-    //    draw_scatter_plot(data, options);
-    //});
+    <script src="{% static 'home/js/alignment.js' %}"></script>
+    <!-- <script src="{% static 'home/js/scatter_plot.js' %}"></script> -->
+    <script src="{% static 'home/js/saveSvgAsPng.js' %}"></script>
+    <script type="text/javascript" charset="utf-8">
+        $('cutoff-val').keydown( function (event) {
+            if (event.keyCode == 13) {
+                var cutoff = parseInt($('#cutoff-val').val());
+                ApplyCutoff(cutoff);
+            };
+        });
+        $('#score-button').click( function () {
+            var cutoff = $('#cutoff-val').val();
+            if (cutoff == '') {
+                cutoff = 40;
+            }
+            window.location.href = 'render_signature_match_scores/'+cutoff;
+        });
+        //$(window).load(function () {
+        //    draw_scatter_plot(data, options);
+        //});
 
-</script>
+    </script>
 {% endblock %}
