+++ seqsign/templates/signature_left_panel.html
@@ -1,0 +1,66 @@
+<table>
+    <!-- Sequence signature -->
+
+    <!-- Empty row for subheader -->
+    <tr>
+        <td class="ali-td ali-td-header-row"></td>
+    </tr>
+
+    <!-- GENERIC NUMBERS -->
+
+    <!-- segment line -->
+    <tr>
+        <td class="ali-td ali-td-first-col"></td>
+    </tr>
+
+    <!-- numbering scheme lines -->
+    {% for ns in a_pos.numbering_schemes %}
+        <tr>
+            <td class="ali-td-generic-num ali-td-first-col">{{ ns.1 }}</td>
+        </tr>
+    {% endfor %}
+    <!-- FEATURE CONSENSUS -->
+    <tr>
+        <td class="ali-td-header-row ali-td-first-col">
+            SIGNATURE CONSENSUS
+        </td>
+    </tr>
+    <!-- Second row for consensus values -->
+    <!-- <tr>
+        <td class="ali-td-header-row ali-td-first-col"></td>
+    </tr> -->
+    <!-- Button collapsing the Residue properties section -->
+    <tr>
+        <td class="ali-td-header-row ali-td-first-col">
+            <button type="button" class="btn btn-info ali-td" data-toggle="collapse" data-target="#signature,#signature-main">+</button>
+        </td>
+    </tr>
+
+    <tbody class='collapse' id='signature'>
+
+    <!-- FEATURE SIGNATURE -->
+
+        <!-- feature frequency -->
+        <tr>
+            <td class="ali-td-header-row ali-td-first-col">
+                RESIDUE PROPERTIES
+            </td>
+        </tr>
+        {% for feature in a_pos.features %}
+            <tr>
+                <td class="ali-td ali-td-first-col">
+                    {{ feature }}
+                </td>
+            </tr>
+        {% endfor %}
+
+        <!-- GENERIC NUMBERS -->
+
+        <!-- numbering scheme lines -->
+        {% for ns in a_pos.numbering_schemes %}
+            <tr>
+                <td class="ali-td-generic-num ali-td-first-col">{{ ns.1 }}</td>
+            </tr>
+        {% endfor %}
+    </tbody>
+</table>
