+++ seqsign/templates/signature_right_panel.html
@@ -7,43 +7,67 @@
             <td id='anchor' class="ali-td">&nbsp;</td>
         {% endfor %}
     </tr>
-    <!-- Sequence signature -->
-    <!-- Collapse button -->
-    <tr>
-        <td class="ali-td-anchor"></td>
-    </tr>
+    <!-- SEQUENCE SIGNATURE -->
     <tr>
         <td class="ali-td-header-row" colspan="{{ num_residue_columns }}">SEQUENCE SIGNATURE</td>
     </tr>
-    <tbody id='signature-main'>
 
     <!-- GENERIC NUMBERS -->
 
-        <!-- segment title line -->
-        <tr id='segments'>
-            {% for s, num in common_segments.items %}
-                <td id='segment-{{ s }}' class="ali-td ali-td-segment-title" colspan="{{ num|length }}">{{ s }}</td>
-                <td id='anchor' class="ali-td-anchor">&nbsp;</td>
-            {% endfor %}
-        </tr>
-
-        <!-- generic number lines -->
-
-        {% for ns, segments in common_generic_numbers.items %}
-        <tr id='gns'>
-            {% for s, num in segments.items %}
-                {% for n, dn in num.items %}
-                    {% if 'Common G-alpha numbering scheme' in a_neg.numbering_schemes.0 %}
-                        <td id='gn-{{ s }}' class="ali-td-generic-num">{{ dn|make_list|slice:'2:'|join:''}}</td>
-                    {% else %}
-                        <td id='gn-{{ s }}' class="ali-td-generic-num">{{ dn|safe }}</td>
-                    {% endif %}
-                {% endfor %}
-                <td class="ali-td-anchor">&nbsp;</td>
-            {% endfor %}
-        </tr>
+    <!-- segment title line -->
+    <tr id='segments'>
+        {% for s, num in common_segments.items %}
+            <td id='segment-{{ s }}' class="ali-td ali-td-segment-title" colspan="{{ num|length }}">{{ s }}</td>
+            <td id='anchor' class="ali-td-anchor">&nbsp;</td>
         {% endfor %}
+    </tr>
 
+    <!-- generic number lines -->
+
+    {% for ns, segments in common_generic_numbers.items %}
+    <tr id='gns'>
+        {% for s, num in segments.items %}
+            {% for n, dn in num.items %}
+                {% if 'Common G-alpha numbering scheme' in a_neg.numbering_schemes.0 %}
+                    <td id='gn-{{ s }}' class="ali-td-generic-num">{{ dn|make_list|slice:'2:'|join:''}}</td>
+                {% else %}
+                    <td id='gn-{{ s }}' class="ali-td-generic-num">{{ dn|safe }}</td>
+                {% endif %}
+            {% endfor %}
+            <td class="ali-td-anchor">&nbsp;</td>
+        {% endfor %}
+    </tr>
+    {% endfor %}
+    <!-- FEATURE CONSENSUS -->
+
+    <!-- Feature names -->
+    <tr>
+        {% for segment, cons in signature_consensus.items %}
+            {% for pos in cons %}
+                <td class="ali-td-stat-summary ali-residue">
+                    <div data-toggle="tooltip" data-placement="top" data-html="true" title="{{ pos.1|safe }}">
+                        {{ pos.0 }}
+                    </div>
+                </td>
+            {% endfor %}
+            <td class="ali-td-anchor">&nbsp;</td>
+        {% endfor %}
+    </tr>
+    <!-- Consensus values -->
+    <tr>
+        {% for segment, cons in signature_consensus.items %}
+            {% for pos in cons %}
+                <td class="ali-td-stat-summary ali-residue res-cons-{{ pos.3 }}">{{ pos.2 }}</td>
+            {% endfor %}
+            <td class="ali-td-anchor">&nbsp;</td>
+        {% endfor %}
+    </tr>
+
+    <!-- Collapse button -->
+    <!-- <tr>
+        <td class="ali-td-anchor"></td>
+    </tr> -->
+    <tbody class='collapse' id='signature-main'>
     <!-- FEATURE SIGNATURE -->
 
         <!-- feature frequency -->
@@ -65,30 +89,6 @@
             </tr>
         {% endfor %}
 
-    <!-- FEATURE CONSENSUS -->
-
-        <!-- Feature names -->
-        <tr>
-            {% for segment, cons in signature_consensus.items %}
-                {% for pos in cons %}
-                    <td class="ali-td-stat-summary ali-residue">
-                        <div data-toggle="tooltip" data-placement="top" data-html="true" title="{{ pos.1|safe }}">
-                            {{ pos.0 }}
-                        </div>
-                    </td>
-                {% endfor %}
-                <td class="ali-td-anchor">&nbsp;</td>
-            {% endfor %}
-        </tr>
-        <!-- Consensus values -->
-        <tr>
-            {% for segment, cons in signature_consensus.items %}
-                {% for pos in cons %}
-                    <td class="ali-td-stat-summary ali-residue res-cons-{{ pos.3 }}">{{ pos.2 }}</td>
-                {% endfor %}
-                <td class="ali-td-anchor">&nbsp;</td>
-            {% endfor %}
-        </tr>
         <!-- GENERIC NUMBERS -->
 
         <!-- generic number lines -->
